requirements.txt
adalflow==1.0.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
# Editable install with no version control (doc_agent==0.1.0)
-e /home/mathisdumont/DID.SIMO.AI_AGENT_EDITOR/documentation_projet
docopt==0.6.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.14
fastjsonschema==2.21.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langcodes==3.5.0
langid==1.1.6
langsmith==0.4.6
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.8.2
mistune==3.1.3
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
ollama==0.5.1
openai==1.93.0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.47.0
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarg==0.1.9
yarl==1.20.1
zstandard==0.23.0
